Add files via upload
There was something wrong with the files uploaded so will amend the previous commit to now include the things worked by Leo in the correct folders.

Author: Bryan Jara Miranda (bjaram09)
