+++ src/main/java/com/bryanjara/proyectotienda/controllers/ProductoController.java
@@ -1,0 +1,4 @@
+package com.bryanjara.proyectotienda.controllers;
+
+public class ProductoController {
+}
