+++ src/main/java/com/bryanjara/proyectotienda/views/ViewRegistrarProducto.java
@@ -1,0 +1,31 @@
+package com.bryanjara.proyectotienda.views;
+
+import javax.swing.*;
+
+public class ViewRegistrarProducto extends JFrame {
+    private JTextField inputNombre;
+    private JComboBox dropdownCategoria;
+    private JTextField inputPrecio;
+    private JTextField peso;
+    private String inputDimensiones;
+    private JPanel panelImagenes;
+    private JTextField inputDescripcion;
+    private JTextField inputInventarioDisponible;
+    private JComboBox dropdownVendedor;
+
+    private JLabel labelNombre;
+    private JLabel labelCategoria;
+    private JLabel labelPrecio;
+    private JLabel labelPeso;
+    private JLabel labelDimensiones;
+    private JLabel labelImagenes;
+    private JLabel labelDescripcion;
+    private JLabel labelInventario;
+    private JLabel labelVendedor;
+
+    public ViewRegistrarProducto(){
+        
+    }
+
+
+}
